pyproject.toml
[project]
name = "urlparsing"
version = "0.1.0"
description = ""
authors = [
    {name = "andy7",email = "vinandy4718@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-polymorphic (>=4.1.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-rest-polymorphic (>=0.1.10,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
